+++ packages/jupyterlab-lsp/src/components/statusbar.tsx
@@ -20,19 +20,21 @@ import {
 import {
   caretDownIcon,
   caretUpIcon,
-  LabIcon,
-  stopIcon,
   circleEmptyIcon,
-  circleIcon
+  circleIcon,
+  LabIcon,
+  stopIcon
 } from '@jupyterlab/ui-components';
 import { WidgetAdapter } from '../adapters/adapter';
 import { collect_documents, VirtualDocument } from '../virtual/document';
 import { LSPConnection } from '../connection';
 import { DocumentConnectionManager } from '../connection_manager';
 import { ILanguageServerManager, ILSPAdapterManager } from '../tokens';
-import { IDocumentWidget } from '@jupyterlab/docregistry';
-import { codeCheckIcon } from '../index';
+import { DocumentRegistry, IDocumentWidget } from '@jupyterlab/docregistry';
 import { DocumentLocator } from './utils';
+import { INotebookModel, NotebookPanel } from '@jupyterlab/notebook';
+import { LanguageServerManager } from '../manager';
+import { codeCheckIcon, codeClockIcon, codeWarningIcon } from './icons';
 
 interface IServerStatusProps {
   server: SCHEMA.LanguageServerSession;
@@ -235,19 +237,53 @@ class LSPPopup extends VDomRenderer<LSPStatus.Model> {
   }
 }
 
+const SELECTED_CLASS = 'jp-mod-selected';
+
 /**
  * StatusBar item.
  */
 export class LSPStatus extends VDomRenderer<LSPStatus.Model> {
   protected _popup: Popup = null;
+  private interactiveStateObserver: MutationObserver;
+
   /**
    * Construct a new VDomRenderer for the status item.
    */
-  constructor(widget_manager: ILSPAdapterManager) {
+  constructor(
+    widget_manager: ILSPAdapterManager,
+    protected displayText: boolean = true
+  ) {
     super(new LSPStatus.Model(widget_manager));
     this.addClass(interactiveItem);
     this.addClass('lsp-statusbar-item');
     this.title.caption = 'LSP status';
+
+    // add human-readable (and stable) class name reflecting otherwise obfuscated typestyle interactiveItem
+    this.interactiveStateObserver = new MutationObserver(() => {
+      const has_selected = this.node.classList.contains(SELECTED_CLASS);
+      if (!this.node.classList.contains(interactiveItem)) {
+        if (!has_selected) {
+          this.addClass(SELECTED_CLASS);
+        }
+      } else {
+        if (has_selected) {
+          this.removeClass(SELECTED_CLASS);
+        }
+      }
+    });
+  }
+
+  protected onAfterAttach(msg: any) {
+    super.onAfterAttach(msg);
+    this.interactiveStateObserver.observe(this.node, {
+      attributes: true,
+      attributeFilter: ['class']
+    });
+  }
+
+  protected onBeforeDetach(msg: any) {
+    super.onBeforeDetach(msg);
+    this.interactiveStateObserver.disconnect();
   }
 
   /**
@@ -259,16 +295,19 @@ export class LSPStatus extends VDomRenderer<LSPStatus.Model> {
     }
     return (
       <GroupItem
-        spacing={2}
+        spacing={this.displayText ? 2 : 0}
         title={this.model.long_message}
         onClick={this.handleClick}
+        className={'lsp-status-group'}
       >
         <this.model.status_icon.react
           top={'2px'}
           kind={'statusBar'}
           title={'LSP Code Intelligence'}
         />
-        <TextItem source={this.model.short_message} />
+        {this.displayText ? (
+          <TextItem source={this.model.short_message} />
+        ) : null}
         <TextItem source={this.model.feature_message} />
       </GroupItem>
     );
@@ -284,6 +323,44 @@ export class LSPStatus extends VDomRenderer<LSPStatus.Model> {
       align: 'left'
     });
   };
+}
+
+export class StatusButtonExtension
+  implements DocumentRegistry.IWidgetExtension<NotebookPanel, INotebookModel> {
+  constructor(
+    private options: {
+      language_server_manager: LanguageServerManager;
+      connection_manager: DocumentConnectionManager;
+      adapter_manager: ILSPAdapterManager;
+    }
+  ) {}
+
+  /**
+   * For statusbar registration and for internal use.
+   */
+  createItem(displayText: boolean = true): LSPStatus {
+    const status_bar_item = new LSPStatus(
+      this.options.adapter_manager,
+      displayText
+    );
+    status_bar_item.model.language_server_manager = this.options.language_server_manager;
+    status_bar_item.model.connection_manager = this.options.connection_manager;
+    return status_bar_item;
+  }
+
+  /**
+   * For registration on notebook panels.
+   */
+  createNew(
+    panel: NotebookPanel,
+    context: DocumentRegistry.IContext<INotebookModel>
+  ): LSPStatus {
+    const item = this.createItem(false);
+    item.addClass('jp-ToolbarButton');
+    panel.toolbar.insertAfter('spacer', 'LSPStatus', item);
+
+    return item;
+  }
 }
 
 type StatusCode =
@@ -317,6 +394,14 @@ const classByStatus: StatusIconClass = {
   initializing: 'preparing',
   initialized_but_some_missing: 'ready',
   connecting: 'preparing'
+};
+
+const iconByStatus: Record<StatusCode, LabIcon> = {
+  waiting: codeClockIcon,
+  initialized: codeCheckIcon,
+  initializing: codeClockIcon,
+  initialized_but_some_missing: codeWarningIcon,
+  connecting: codeClockIcon
 };
 
 const shortMessageByStatus: StatusMap = {
@@ -509,7 +594,7 @@ export namespace LSPStatus {
       if (!this.adapter) {
         return stopIcon;
       }
-      return codeCheckIcon.bindprops({
+      return iconByStatus[this.status.status].bindprops({
         className: 'lsp-status-icon ' + classByStatus[this.status.status]
       });
     }