+++ CHANGELOG.md
@@ -2,11 +2,17 @@
 
 ### `@krassowski/jupyterlab-lsp 3.0.1` (unreleased)
 
+- features
+
+  - make the extension work with `jupyterlab-classic` - experimental, not all features are functional yet ([#465])
+
 - bug fixes:
 
   - fix completions with R double and triple colon prefix ([#449])
+  - fix contrast on status icon when status item is active ([#465])
 
 [#449]: https://github.com/krassowski/jupyterlab-lsp/pull/449
+[#465]: https://github.com/krassowski/jupyterlab-lsp/pull/465
 
 ### `jupyter-lsp 1.0.1` (unreleased)